Check the ambiance: Light sensor, Backlight and Fade. Yeah so the backlight comes on for n millis, then fades out unless another press is given. Ambient sensor tests on backlight intensity. Button debounce changed. It still sucks. Menu subtext changed

+++ RebbleOS/display.h
@@ -66,7 +66,7 @@ typedef struct {
     
     //state
     uint8_t BacklightEnabled;
-    uint8_t Brightness;
+    uint16_t Brightness;
     uint8_t PowerOn;
     uint8_t State; // busy etc
     uint8_t DisplayMode; // bootloader or full
